refactor(site): Regroupe les composants du footer.
