+++ site/source/components/layout/Footer/LegalNotice.tsx
